feat: CRUD PACIENTE Y ODONTOLOGO, MANEJO DE EXCEPCIONES

+++ src/main/java/com/proyectoIntegrador/PrjInt_GustavoContreras_BrendaCarbajal/servicio/impl/PacienteServicioImpl.java
@@ -1,21 +1,19 @@
 package com.proyectoIntegrador.PrjInt_GustavoContreras_BrendaCarbajal.servicio.impl;
 
 import com.proyectoIntegrador.PrjInt_GustavoContreras_BrendaCarbajal.entity.Paciente;
+import com.proyectoIntegrador.PrjInt_GustavoContreras_BrendaCarbajal.exception.ResourceNotFoundException;
 import com.proyectoIntegrador.PrjInt_GustavoContreras_BrendaCarbajal.repository.IPacienteRepository;
 import com.proyectoIntegrador.PrjInt_GustavoContreras_BrendaCarbajal.servicio.IPacienteServicio;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
-public class PacienteServicioImpl  implements IPacienteServicio {
+public class PacienteServicioImpl implements IPacienteServicio {
 
     @Autowired
     private IPacienteRepository iPacienteRepository;
-
-
 
     @Override
     public Paciente guardar(Paciente paciente) {
@@ -24,13 +22,8 @@ public class PacienteServicioImpl  implements IPacienteServicio {
 
     @Override
     public Paciente buscarPorId(Long id) {
-        Optional<Paciente> pacienteBuscado = iPacienteRepository.findById(id);
-        if(!pacienteBuscado.isPresent())  {
-            return null;
-        } else {
-            return pacienteBuscado.get();
-        }
-
+        return iPacienteRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Paciente no encontrado con id: " + id));
     }
 
     @Override
@@ -40,13 +33,17 @@ public class PacienteServicioImpl  implements IPacienteServicio {
 
     @Override
     public void actualizar(Paciente paciente) {
+        if (!iPacienteRepository.existsById(paciente.getId())) {
+            throw new ResourceNotFoundException("Paciente no encontrado con id: " + paciente.getId());
+        }
         iPacienteRepository.save(paciente);
     }
 
     @Override
     public void eliminar(Long id) {
+        if (!iPacienteRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Paciente no encontrado con id: " + id);
+        }
         iPacienteRepository.deleteById(id);
     }
-
-
 }
